refactor(`common/model`): Change the folder to `entities`
