+++ lib/common/domain/entities/weather.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_training/common/models/weather_condition.dart';
+import 'package:flutter_training/common/domain/entities/weather_condition.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'weather.freezed.dart';
