+++ lib/common/domain/entities/weather.freezed.dart
