+++ lib/common/domain/entities/weather_condition.dart
