+++ lib/feature/day_weather/components/weather_info.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_training/common/components/weather_icon.dart';
-import 'package:flutter_training/common/models/weather.dart';
+import 'package:flutter_training/common/domain/entities/weather.dart';
 import 'package:flutter_training/feature/day_weather/components/temperature_text.dart';
 
 class WeatherInfo extends StatelessWidget {
