+++ lib/feature/day_weather/repository.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
 import 'package:flutter_training/common/utils/result.dart';
-import 'package:flutter_training/common/models/weather.dart';
-import 'package:flutter_training/common/models/weather_condition.dart';
+import 'package:flutter_training/common/domain/entities/weather.dart';
+import 'package:flutter_training/common/domain/entities/weather_condition.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:simple_logger/simple_logger.dart';
 import 'package:yumemi_weather/yumemi_weather.dart';
